# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Booking API",
        "description": "A Laravel-based API for booking services.",
        "contact": {
            "email": "smazoomder@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get a list of available services",
                "description": "Retrieves a paginated list of services with optional pagination query parameters. The results are cached for performance.",
                "operationId": "getServices",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "The number of services per page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Services retrieved from cache."
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/services?page=1"
                                                },
                                                "last_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/services?page=2"
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/services?page=2",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Service"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error for pagination parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid pagination parameters."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Service": {
                "required": [
                    "id",
                    "name",
                    "category",
                    "price",
                    "description"
                ],
                "properties": {
                    "id": {
                        "description": "Service ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Service name",
                        "type": "string"
                    },
                    "category": {
                        "description": "Service category",
                        "type": "string"
                    },
                    "price": {
                        "description": "Service price",
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "description": "Service description",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a valid bearer token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}